fix(linux): Fix debian postinst script
The ibus-keyman post-installation script failed, causing the entire
installation to fail. This change removes `set -e` so that we try
to run the rest of the commands even when one command should fail.
Also accounts for ibus-daemon being started with the full path,
and start ibus-daemon with fill path as well.