chore(developer): cleanup API interfaces and types for kmcmplib
Introduces kmcmplibapi.h which declares all external functions and types
for kmcmplib, and updates kmcmplib and kmcmpdll to use these
declarations consistently.

Note that kmcmp_CompileKeyboardToBuffer declares a void* instead of
PFILE_KEYBOARD, because this is a large and complex type defined in both
kmcmpdll and kmcmplib. In the future, when kmcmpdll existing legacy
compiler is removed, we can make the kmcmplib type more public.