fix(windows): use right instead of left alignment
The fixed layout of the popup means we need use specified
starting point for the child div that contains the add language
and close button. However, it needs to be right justfified
instead of left justified to allow for the variable sized
buttons. The buttons will be longer based on the locale
text chosen.