chore(linux): Fix installation build step on TC
When building on TC we install into a temporary directory. This
broke with the recent introduction of build.sh because that
now calls `make install` whereas previously we explicitly called
`make install-temp` on TC. Instead of adding a `install-temp`
action to `build.sh` we now check in the Makefile if sudo is set.