fix(developer): compiler needs to support loading .kvk files
Fixes #9150.

Legacy .kmn keyboards can refer to a binary .kvk file. kmc needs to be
able to load these as well.

Note: there was a short period of time in which .kvk files were either
source or binary format. We moved to .kvk always being binary, and .kvks
always being source (xml), and so IMO we don't need to include support
for mismatched file formats.