chore(linux): Remove Kinetic from GHA
Don't build Ubuntu 22.10 Kinetic with packaging GHA.

Also prepare for Ubuntu 23.10 Mantic - we can't currently
build that because they are trying to get ibus 1.5.29-beta
to work on so there are often new packages. We wait until that
stabilized before we're providing a new patched ibus version
at which point we can enable building for mantic.
See also #9398.