fix(windows): allow qr box to grow to edge of default
Allow QR box to grow if the caption box is bigger than
the english text. Still limit this to 130px if it grows
further than this then add a vertical scroll for the caption.

In testing no language actually causes the scroll bar to be
displayed. However using dummy text to test the overflow
shows that it will handle this should the translation ever
reach this length.