feat(developer): show an INFO message when warnings have failed a build
Fixes #9610.

If a build fails when "treat warnings as errors" is on, but there are
only warnings and no error messages, it can be confusing for the
developer. Adds the following `INFO_WarningsHaveFailedBuild` message:

`The build failed because option "treat warnings as errors" is enabled
and there are one or more warnings.`